+++ R/trelliscope-data.R
@@ -127,3 +127,5 @@ NULL
 #' @usage data(batting)
 #' @format A data frame with 1406 rows and 22 variables
 NULL
+
+